added element property to document object
