+++ docx/blkcntnr.py
@@ -11,7 +11,7 @@ from __future__ import absolute_import, division, print_function, unicode_litera
 from docx.oxml.table import CT_Tbl
 from docx.shared import Parented
 from docx.text.paragraph import Paragraph
-
+from docx.api import element
 
 class BlockItemContainer(Parented):
     """Base class for proxy objects that can contain block items.
@@ -82,13 +82,3 @@ class BlockItemContainer(Parented):
         """
         return Paragraph(self._element.add_p(), self)
 
-
-def element(element, part):
-    if str(type(element)) == "<class 'docx.oxml.text.paragraph.CT_P'>":
-        return Paragraph(element, element.getparent())
-    elif str(type(element)) == "<class 'docx.oxml.table.CT_Tbl'>":
-        from .table import Table
-        return Table(element, element.getparent())
-    elif str(type(element)) == "<class 'docx.oxml.section.CT_SectPr'>":
-        from .section import Section
-        return Section(element, part)
